#22: Rename classes and interface for better abstraction
