+++ my-webxr-app/src/repository/Repository.tsx
@@ -1,7 +1,9 @@
 import {DataPoint} from "./DataPoint.tsx";
-import {Column} from "./IndexDbRepository.tsx";
+import {Column} from "./DbRepository.tsx";
 
-export interface ColumnRepository{
+// The repository interface defines operations that can be
+// done to/on the db.
+export interface Repository {
     getPoints: (qualifyingPointOnly : boolean) => Promise<Array<DataPoint>>;
     addColumn: (column : Column) => void;
     addColumnBulk: (columns : Array<Column>) => void;
